Reorganize systematics tests
